BJS2-27037 refactoring resolve conflicts
